super a'can groundwork / splitting into logical per-chip parts
